Brought in Level DB storage. (#60)
Updated MapDB to store CBOR serialized objects, not java
serialized objects.

Ensured that the SodiumEncryptedPayload will store the mapping of
public key to combined key cleanly. Added string constructor to
SodiumPublicKey will be based on on the string encoded version (for the
json map key version of the data).

Added type parameter to the MapDBStorage to help with
serialization/deserialization.

Added JDK8 module object mapper to help with the handling of optionals
to and from JSON and CBOR.

Actually use the database from the config.